requirements.txt
Django==1.8.2
Pillow==2.8.2
djangorestframework==3.1.3
drf-nested-routers==0.9.0
